Fractional position sizing, and the crypto verdict it unlocked
risk/sizing.py floored every quantity to a whole unit, so a Rs 250 risk
budget against BTC's ~Rs 77 lakh price sized to 0. BTC, ETH and BNB had
therefore never been backtested even once - every signal was silently
skipped as "size 0", and the earlier "majors" result was really a
verdict on cheap altcoins only.

Sizing now quantizes to a qty_step carried on Market/BacktestConfig/
SizeInputs, plus CoinDCX's flat Rs 100 min_notional (both fetched live
from markets_details). NSE keeps step 1.0 and is unchanged: a hypothesis
property test asserts it reproduces the old math.floor path exactly, and
the 5 ledger-verified golden contract notes are untouched.

With the majors actually included the answer is clear and negative: 253
trades over 2018-2026, gross P&L -Rs 5,198 (no edge before costs) and
-Rs 20,272 of costs on top. A round trip costs 1.47% of position value,
68% of which is the 1% Section 194S TDS on the sell leg - levied on sell
value win or lose, so a ~3-day hold must gain 1.47% just to break even.
BTC and ETH were the worst performers, not the best.

+++ src/tradedesk/engine/filters.py
@@ -30,7 +30,7 @@ class FilterResult:
     net_rr_t2: float | None = None
 
 
-def net_rr(sig: Signal, qty: int, costs: CostModel) -> tuple[float | None, float | None]:
+def net_rr(sig: Signal, qty: float, costs: CostModel) -> tuple[float | None, float | None]:
     if qty <= 0:
         return None, None
     common = dict(trade_type=TradeType.DELIVERY, qty=qty, entry=_d(sig.trigger), stop=_d(sig.stop))
@@ -46,7 +46,7 @@ def _d(x: float) -> Decimal:
 def apply_filters(
     sig: Signal,
     *,
-    qty: int,
+    qty: float,
     atr_pct: float | None,
     avg_turnover: float | None,
     regime: str | None,
